reformatting. Moved dkb specific content in separate file module
