+++ pynance/textimporter.py
@@ -1,9 +1,6 @@
-import csv
-import datetime
 import re
 
 import pandas as pd
-
 import numpy as np
 
 
@@ -83,6 +80,31 @@ def read_csv(filepath_or_buffer, description):
                                                   final_total_balance)
 
     return new_df
+
+
+def amounts_to_balances(amounts, final_balance):
+    """
+    Gives a list of balances after each transaction
+    Calculated backwards, starting with the given final balance
+
+    PARAMS:
+    -------
+    amounts : iterable of float
+        amount transferred for each transaction
+    final_balance : float
+        total value of the balance after the last transaction
+
+    RETURNS:
+    --------
+    list of float
+        values of the total balance after each transaction
+    """
+    if len(amounts) == 1:
+        return [final_balance]
+    else:
+        return amounts_to_balances(amounts[:-1],
+                                   final_balance-amounts[-1]) \
+            + [final_balance]
 
 
 class CsvFileDescription():
@@ -198,111 +220,3 @@ COLUMNS = {
     "category": str,
     "tags": str,
     "origin": str}
-
-# DKB definitions
-# for most solutions, see
-# https://github.com/hamvocke/dkb2homebank/blob/master/dkb2homebank.py
-# (MIT licenced)
-
-
-class DKBFormatters():
-    """
-        contains all the static functions needed to convert a string
-        as given from a DKB CSV file to all the types needed in COLUMNS
-    """
-
-    @classmethod
-    def to_datetime64(cls, datestring):
-        # TODO: there is probably a better way to convert it
-        date = datetime.datetime.strptime(datestring, "%d.%m.%Y")
-        return np.datetime64(date.strftime('%Y-%m-%d'))
-
-    @classmethod
-    def to_string(cls, anystring):
-        # remove trailing whitespace
-        return anystring.strip()
-
-    @classmethod
-    def to_float64(cls, numberstring):
-        if not numberstring:
-            return np.nan
-        else:
-            # TODO: use some builtin package, maybe locale for this
-            return float(numberstring.replace(".", "").replace(",", "."))
-
-    @classmethod
-    def formatter_map(cls):
-        return {
-            np.datetime64: cls.to_datetime64,
-            str: cls.to_string,
-            np.float64: cls.to_float64
-        }
-
-
-class DKBCsvDialect(csv.Dialect):
-    """
-        Statically defines what a DKBCash CSV-file looks like
-    """
-    delimiter = ';'
-    quotechar = '"'
-    doublequote = True
-    skipinitialspace = False
-    lineterminator = '\r\n'
-    quoting = csv.QUOTE_MINIMAL
-
-
-class SupportedCsvTypes():
-    """
-        Static enumeration of all the supported CSV file types, to be used
-        with :func:`~textimporter.read_csv~`
-    """
-    DKBCash = CsvFileDescription(
-        column_map={
-            "date": "Wertstellung",
-            "sender_account": "Kontonummer",
-            "text": "Verwendungszweck",
-            "amount": "Betrag (EUR)",
-        },
-        csv_dialect=DKBCsvDialect(),
-        formatters=DKBFormatters.formatter_map(),
-        skiprows=6,
-        encoding="iso-8859-1",
-        total_balance_re_pattern=r'(?<=Kontostand vom \d{2}.\d{2}.\d{4}:";")'
-                                 r'(\d+,\d+)')
-
-    DKBVisa = CsvFileDescription(
-        column_map={
-            "date": "Wertstellung",
-            "text": "Beschreibung",
-            "amount": "Betrag (EUR)",
-        },
-        csv_dialect=DKBCsvDialect(),
-        formatters=DKBFormatters.formatter_map(),
-        skiprows=6,
-        encoding="iso-8859-1",
-        total_balance_re_pattern=r'(?<=Saldo:";")(\d+,\d+)')
-
-
-def amounts_to_balances(amounts, final_balance):
-    """
-    Gives a list of balances after each transaction
-    Calculated backwards, starting with the given final balance
-
-    PARAMS:
-    -------
-    amounts : iterable of float
-        amount transferred for each transaction
-    final_balance : float
-        total value of the balance after the last transaction
-
-    RETURNS:
-    --------
-    list of float
-        values of the total balance after each transaction
-    """
-    if len(amounts) == 1:
-        return [final_balance]
-    else:
-        return amounts_to_balances(amounts[:-1],
-                                   final_balance-amounts[-1]) \
-            + [final_balance]
